resolve more comments
