+++ lib/types/Cargo.toml
@@ -24,7 +24,7 @@ target-lexicon = { version = "0.12.2", default-features = false }
 enumset.workspace = true
 bytecheck = "0.6.8"
 rand = "0.8"
-webc = { workspace = true }
+webc = { workspace = true, default-features = false }
 xxhash-rust = { version = "0.8.8", features = ["xxh64"] }
 sha2 = { version = "0.10" }
 hex = { version = "^0.4" }
