+++ lib/types/src/lib.rs
@@ -58,6 +58,7 @@ mod initializers;
 mod libcalls;
 mod memory;
 mod module;
+mod module_hash;
 mod serialize;
 mod stack;
 mod store_id;
@@ -96,7 +97,8 @@ pub use crate::initializers::{
     OwnedDataInitializer, TableInitializer,
 };
 pub use crate::memory::{Memory32, Memory64, MemorySize};
-pub use crate::module::{ExportsIterator, ImportKey, ImportsIterator, ModuleHash, ModuleInfo};
+pub use crate::module::{ExportsIterator, ImportKey, ImportsIterator, ModuleInfo};
+pub use crate::module_hash::ModuleHash;
 pub use crate::units::{
     Bytes, PageCountOutOfRange, Pages, WASM_MAX_PAGES, WASM_MIN_PAGES, WASM_PAGE_SIZE,
 };
